Fix global_mean truncation in classifier and features to match ESP
classifier.py: NIGHT threshold comparison now uses int truncation —
a float global_mean of 70.5 would differ from ESP's integer 70 and
cause a spurious NIGHT/not-NIGHT parity mismatch.
features.py: global_mean output also truncated for consistency.

+++ src/cloud-check/cloud_check/features.py
@@ -52,5 +52,5 @@ def extract_tile_features(
     return {
         "mean": mean,
         "std": std,
-        "global_mean": float(f.mean()),
+        "global_mean": int(f.mean()),  # truncate, matches ESP integer division
     }
